requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
urllib3==1.26.7
